Move content to to File.kt. Also remove deleteAllFiles, we have deleteRecursively now.
